Add read/write in list functions
Add functionality to handle data from a csv file in terms of
a list of ordered dictionaries.

TODO: this class is a wrapper around a csv file. As such, it
should only be responsible for reading/writing those files. There
should be now self.word_list internal to the class. It should
return lists of OrderedDicts for other classes to use.