Alter tests to ensure new correct behavior
Alter tests to reflect that data is now being stored in lists
of ordered dictionaries, (as that's what csvDictReader returns)
rather than list of lists.

Accomplish a new set of tests by using the new CsvFileHandle class
to handle writing to files, changing list indexes to dict keys,
and placing lists of ordered dicts as the expected data format.